move aggregate_to_country to custom_aggregations. add reporting_ratio_required_to_aggregate param to break import loop.
